refactor: rename return component
