--- src/app/pages/return/return-routing.module.ts
@@ -1,17 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { ReturnPage } from './return.page';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ReturnPage
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule],
-})
-export class ReturnPageRoutingModule {}
