--- src/app/pages/return/return.module.ts
@@ -1,15 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
-
-import {IonicModule} from '@ionic/angular';
-
-import {ReturnPageRoutingModule} from './return-routing.module';
-
-import {ReturnPage} from './return.page';
-
-@NgModule({
-  imports: [CommonModule, FormsModule, IonicModule, ReturnPageRoutingModule],
-  declarations: [ReturnPage],
-})
-export class ReturnPageModule {}
